Split adapter debugging script into common functions
This includes a bug fix in `scripts/lib/debug.ts` (line 29 to 42):
The requests from datasette need to be reformatted first before they can be filtered
because the `match()` function might depend on the content being a string.
